requirements.txt
pyswisseph>=2.10.3
python-dateutil>=2.8.2
pytz>=2024.1
requests>=2.31.0
astroquery>=0.4.7
